feat: tno-1328, tno-1329 - Tone and Topic mappping - changes as per PR feedback

+++ libs/net/kafka/Models/TonePool.cs
@@ -3,7 +3,7 @@ namespace TNO.Kafka.Models;
 public class TonePool
 {
     #region Properties
-    public long Value { get; set; }
+    public int Value { get; set; }
     public string? UserIdentifier { get; set; }
 
     #endregion
@@ -11,7 +11,7 @@ public class TonePool
     #region Constructors
     public TonePool() { }
 
-    public TonePool(long toneValue, string userIdentifier)
+    public TonePool(int toneValue, string userIdentifier)
     {
         this.Value = toneValue;
         this.UserIdentifier = userIdentifier ?? "";
